Factor out layout components to their own files
