+++ package.json
@@ -23,7 +23,7 @@
 		"@sveltejs/adapter-node": "^1.3.1",
 		"@sveltejs/kit": "^1.27.6",
 		"sass": "^1.69.5",
-		"svelte": "^5.0.0-next.6",
+		"svelte": "^5.0.0-next.8",
 		"svelte-preprocess": "^5.1.0",
 		"vite": "^5.0.0"
 	},
